blog_to_voice/requirements.txt
streamlit
python-dotenv
agno
elevenlabs
requests
pydantic